# api/openapi/openapi.yaml
...
openapi: 3.0.3
info:
  title: Imagine API
  version: 0.1.0
  description: OpenAPI specification for the Imagine API (chi-based)
servers:
  - url: http://localhost:7770
    description: Local API
paths:
...
  /jobs:
    get:
      summary: List jobs
      operationId: listJobs
      responses:
        "200":
          description: Jobs list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobListResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: Create/enqueue a job
      operationId: createJob
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobCreateRequest"
      responses:
        "202":
          description: Accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobEnqueueResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /jobs/{id}:
    get:
      summary: Get job detail
      operationId: getJob
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Job detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobInfo"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      summary: Cancel job
      operationId: cancelJob
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: Retry job
      operationId: retryJob
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "501":
          description: Not implemented
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /jobs/scheduler/start:
    post:
      summary: Start job scheduler
      operationId: startScheduler
      responses:
        "200":
          description: Started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /jobs/scheduler/shutdown:
    post:
      summary: Shutdown job scheduler
      operationId: shutdownScheduler
      responses:
        "200":
          description: Shutdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /jobs/types/{type}/stop:
    post:
      summary: Stop a job type
      operationId: stopJobType
      parameters:
        - in: path
          name: type
          required: true
          schema:
            type: string
          description: Job type to stop
      responses:
        "200":
          description: Job type stopped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /jobs/types/{type}/concurrency:
    put:
      summary: Update job type concurrency
      operationId: updateJobTypeConcurrency
      parameters:
        - in: path
          name: type
          required: true
          schema:
            type: string
          description: Job type to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                concurrency:
                  type: integer
                  minimum: 1
                  maximum: 100
              required: [concurrency]
      responses:
        "200":
          description: Concurrency updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    JobInfo:
      type: object
      properties:
        id:
          type: string
        topic:
          type: string
        status:
          type: string
      required: [id, topic, status]
    JobListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/JobInfo"
      required: [items]
    JobCreateRequest:
      type: object
      properties:
        type:
          type: string
          description: "Job type (e.g., thumbnailGeneration, imageProcessing)"
        command:
          type: string
          enum: [all, missing, single]
          description: "Command to execute (all=process all, missing=process missing, single=process one image)"
        image_uid:
          type: string
          description: "Image UID (required when command=single)"
      required: [type]
    JobEnqueueResponse:
      type: object
      properties:
        message:
          type: string
        count:
          type: integer
      required: [message]
...